backend/requirements.txt
Django
python-environ
djangorestframework
psycopg2-binary
django-cors-headers
dj-rest-auth
django-allauth
djangorestframework-simplejwt
djangorestframework-camel-case
Pillow
uvicorn[standard]
